Added workaround for ExtAudioFileSeek
It seems (at least on Leopard) that to get ExtAudioFileSeek working requires that a client data format is set on
the ExtAudioFileRef.  So, set the client format to match the file's data format.

Also, the number of blocks was incorrectly calculated- it was off by one since the range is inclusive.